+++ infrastructure/modules/onpremise-storage/mongodb/mongodb.tf
@@ -28,6 +28,7 @@ resource "kubernetes_deployment" "mongodb" {
         }
       }
       spec {
+        node_selector = var.mongodb.node_selector
         dynamic toleration {
           for_each = (var.mongodb.node_selector != {} ? [
           for index in range(0, length(local.node_selector_keys)) : {