+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Added
 
 * Add HPA on ArmoniK control plane based on CPU and RAM utilization metrics
 * Add node selector on all deployments (ArmoniK and Helm charts for infrastructure)
+* Add logs of metrics exporter, MongoDB, NGINX, and Keda in CloudWatch via Fluent-bit
 
 ## [v2.8.1](https://github.com/aneoconsulting/armonik/tree/v2.8.1) (2022-07-01)